requirements_mysql.txt
# ===========================================
# DEPENDENCIAS PARA PROYECTO MULTI-FRAMEWORK
# Flask + Django + FastAPI + SQLAlchemy + MySQL/MariaDB
# ===========================================

# === FLASK Y EXTENSIONES ===
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==6.0.1
Flask-JWT-Extended==4.7.1

# === DJANGO Y EXTENSIONES ===
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1

# === FASTAPI Y EXTENSIONES ===
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# === SQLALCHEMY Y BASE DE DATOS ===
SQLAlchemy==2.0.43
alembic==1.16.5

# === DRIVERS DE BASE DE DATOS ===
# MySQL/MariaDB
PyMySQL==1.1.2
mysqlclient==2.2.7

# PostgreSQL (opcional)
psycopg2-binary==2.9.10

# === SERIALIZACIÓN Y VALIDACIÓN ===
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
pydantic==2.11.9

# === UTILIDADES ===
python-dotenv==1.1.1

# === DEPENDENCIAS TRANSITIVAS (instaladas automáticamente) ===
# asgiref==3.9.2
# bcrypt==5.0.0
# blinker==1.9.0
# cffi==2.0.0
# click==8.3.0
# colorama==0.4.6
# cryptography==46.0.1
# ecdsa==0.19.1
# greenlet==3.2.4
# h11==0.16.0
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.6
# Mako==1.3.10
# MarkupSafe==3.0.3
# pyasn1==0.6.1
# pycparser==2.23
# pydantic_core==2.33.2
# PyJWT==2.10.1
# rsa==4.9.1
# six==1.17.0
# sniffio==1.3.1
# sqlparse==0.5.3
# starlette==0.48.0
# typing-inspection==0.4.1
# typing_extensions==4.15.0
# tzdata==2025.2
# Werkzeug==3.1.3
